+++ shape/rectangle.go
@@ -1,10 +1,10 @@
 package shape
 
 type Rectangle struct {
-	width  float32
-	height float32
+	width  float64
+	height float64
 }
 
-func (r Rectangle) Area() float32 {
+func (r Rectangle) Area() float64 {
 	return r.width * r.height
 }
